chore(deps): pin jamfplatform-go-sdk to 54d3840; bump min Jamf Pro to 11.29.0
Pin the SDK to commit 54d3840 (v0.9.1-0.20260618090315-54d3840d1399) and
resolve the breaking changes it introduces:

- LAPS settings: the SDK split the single LocalAdminPasswordSettings type into
  LapsSettingsRequestV2 (write) / LapsSettingsResponseV2 (read) and changed the
  duration fields from int64 to int. Update local_admin_password_settings to
  read responses, build requests from a response merge-base, retype the duration
  maps/helpers to int, and add a lapsResponseToRequest converter for the acc-test
  read-then-write-back paths (the SDK no longer lets a read feed a write directly).
- computer_prestage_enrollment: the SDK dropped AuthURL from GetComputerPrestageV3.
  It was a Computed-only server-derived field, so drop the auth_url attribute from
  schema/model/state-builder (provider unshipped; no state-upgrader concern).

Raise ProviderMinJamfProVersion 11.28.0 -> 11.29.0 (and its doc/README/STYLE_GUIDE
references and at-floor test mocks).

+++ internal/resources/pro/settings/local_admin_password_settings/schema_types.go
@@ -28,7 +28,7 @@ const rotationIntervalNever = "Never"
 // UI dropdown exactly (Settings → Computer Management → Security). The server
 // stores arbitrary durations and does not enforce these presets, so the
 // provider's OneOf validator is the sole guard on the configured value.
-var rotationAfterViewingToDuration = map[string]int64{
+var rotationAfterViewingToDuration = map[string]int{
 	"1 hour":   3600,
 	"3 hours":  10800,
 	"12 hours": 43200,
@@ -45,7 +45,7 @@ var validRotationAfterViewingInterval = []string{
 
 // rotationIntervalDurationToValue maps each non-"Never" "Rotation interval"
 // dropdown preset to the duration the server stores for it.
-var rotationIntervalDurationToValue = map[string]int64{
+var rotationIntervalDurationToValue = map[string]int{
 	"7 days":   604800,
 	"30 days":  2592000,
 	"60 days":  5184000,
@@ -92,8 +92,8 @@ func markdownValueList(vals []string) string {
 }
 
 // invertDurationMap builds the duration -> label reverse of a label -> duration map.
-func invertDurationMap(m map[string]int64) map[int64]string {
-	out := make(map[int64]string, len(m))
+func invertDurationMap(m map[string]int) map[int]string {
+	out := make(map[int]string, len(m))
 	for label, dur := range m {
 		out[dur] = label
 	}
